requirements.txt
ultralytics
opencv-python
numpy
ollama
pyserial
python-dotenv
serial
python-escpos
